Refactor the code into its own echo module
