+++ python/semantic_kernel/connectors/ai/echo/__init__.py
@@ -1,0 +1,13 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+from semantic_kernel.connectors.ai.echo.services.echo_text_completion import (
+    EchoTextCompletion,
+)
+from semantic_kernel.connectors.ai.echo.services.echo_chat_completion import (
+    EchoChatCompletion,
+)
+
+__all__ = [
+    "EchoTextCompletion",
+    "EchoChatCompletion",
+]
