+++ python/semantic_kernel/connectors/ai/echo/services/echo_text_completion.py
